+++ src/main/java/org/spongepowered/common/event/tracking/phase/packet/PacketFunction.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.event.tracking.phase.util;
+package org.spongepowered.common.event.tracking.phase.packet;
 
 import com.flowpowered.math.vector.Vector3d;
 import com.google.common.collect.ImmutableList;
@@ -81,7 +81,6 @@ import org.spongepowered.common.event.tracking.CauseTracker;
 import org.spongepowered.common.event.tracking.ItemDropData;
 import org.spongepowered.common.event.tracking.PhaseContext;
 import org.spongepowered.common.event.tracking.TrackingUtil;
-import org.spongepowered.common.event.tracking.phase.PacketPhase;
 import org.spongepowered.common.interfaces.IMixinContainer;
 import org.spongepowered.common.interfaces.network.IMixinNetHandlerPlayServer;
 import org.spongepowered.common.interfaces.world.IMixinWorldServer;
@@ -119,7 +118,7 @@ public interface PacketFunction {
         final CauseTracker causeTracker = mixinWorldServer.getCauseTracker();
         EntityUtil.toMixin(entity).setNotifier(player.getUniqueID());
 
-        if (state == PacketPhase.General.ATTACK_ENTITY) {
+        if (state == GeneralPacketState.ATTACK_ENTITY) {
             context.getCapturedItemsSupplier()
                     .ifPresentAndNotEmpty(items -> {
                         // For destruction, this should be empty, however, some times, it may not be?
@@ -198,7 +197,7 @@ public interface PacketFunction {
             });
 
 
-        } else if (state == PacketPhase.General.INTERACT_ENTITY) {
+        } else if (state == GeneralPacketState.INTERACT_ENTITY) {
             context.getCapturedBlockSupplier()
                     .ifPresentAndNotEmpty(blocks -> {
                         final PrettyPrinter printer = new PrettyPrinter(80);
@@ -250,7 +249,7 @@ public interface PacketFunction {
                         printer.trace(System.err);
                     });
 
-        } else if (state == PacketPhase.General.INTERACT_AT_ENTITY) {
+        } else if (state == GeneralPacketState.INTERACT_AT_ENTITY) {
             context.getCapturedEntitySupplier().ifPresentAndNotEmpty(entities -> {
                 final Cause cause = Cause.source(EntitySpawnCause.builder()
                         .entity((Player) player)
@@ -365,7 +364,7 @@ public interface PacketFunction {
                 .orElse(null);
         final ItemStackSnapshot usedSnapshot = ItemStackUtil.snapshotOf(usedStack);
         final Entity spongePlayer = EntityUtil.fromNative(player);
-        if (state == PacketPhase.Inventory.DROP_ITEM) {
+        if (state == InventoryPacketState.DROP_ITEM) {
             context.getCapturedBlockSupplier()
                     .ifPresentAndNotEmpty(blocks ->
                             TrackingUtil.processBlockCaptures(blocks, causeTracker, state, context)
@@ -395,7 +394,7 @@ public interface PacketFunction {
 
 
                     });
-        } else if (state == PacketPhase.Inventory.DROP_INVENTORY) {
+        } else if (state == InventoryPacketState.DROP_INVENTORY) {
 
             context.getCapturedBlockSupplier()
                     .ifPresentAndNotEmpty(blocks -> TrackingUtil.processBlockCaptures(blocks, causeTracker, state, context));
@@ -421,7 +420,7 @@ public interface PacketFunction {
                         }
                     });
 
-        } else if (state == PacketPhase.General.INTERACTION) {
+        } else if (state == GeneralPacketState.INTERACTION) {
             context.getCapturedBlockSupplier()
                     .ifPresentAndNotEmpty(blocks ->
                             TrackingUtil.processBlockCaptures(blocks, causeTracker, state, context)
@@ -624,9 +623,9 @@ public interface PacketFunction {
         }
         final Cause cause = Cause.of(NamedCause.source(player), NamedCause.of("Container", openContainer));
         final InteractInventoryEvent inventoryEvent;
-        if (state instanceof PacketPhase.Inventory) {
+        if (state instanceof InventoryPacketState) {
             inventoryEvent =
-                    ((PacketPhase.Inventory) state)
+                    ((InventoryPacketState) state)
                             .createInventoryEvent(player, ContainerUtil.fromNative(openContainer), transaction, slotTransactions, capturedItems,
                                     cause, usedButton);
         } else {
@@ -698,7 +697,7 @@ public interface PacketFunction {
 
     });
     PacketFunction PLACE_BLOCK = (packet, state, player, context) -> {
-        if (state == PacketPhase.General.INVALID) { // This basically is an out of world place, and nothing should occur here.
+        if (state == GeneralPacketState.INVALID) { // This basically is an out of world place, and nothing should occur here.
             return;
         }
         final IMixinWorldServer mixinWorld = (IMixinWorldServer) player.worldObj;
@@ -861,7 +860,7 @@ public interface PacketFunction {
         SpongeImpl.postEvent(event);
     });
     PacketFunction CLIENT_STATUS = ((packet, state, player, context) -> {
-        if (state == PacketPhase.Inventory.OPEN_INVENTORY) {
+        if (state == InventoryPacketState.OPEN_INVENTORY) {
             final ItemStackSnapshot lastCursor = context.firstNamed(InternalNamedCauses.Packet.CURSOR, ItemStackSnapshot.class)
                     .orElseThrow(TrackingUtil.throwWithContext("Expected a cursor item stack, but had nothing!", context));
             final ItemStackSnapshot newCursor = ItemStackUtil.snapshotOf(player.inventory.getItemStack());
@@ -1028,6 +1027,6 @@ public interface PacketFunction {
 
     PacketFunction HANDLED_EXTERNALLY = UNKONWN_PACKET;
 
-    void unwind(Packet<?> packet, PacketPhase.IPacketState state, EntityPlayerMP player, PhaseContext context);
+    void unwind(Packet<?> packet, IPacketState state, EntityPlayerMP player, PhaseContext context);
 
 }
