+++ src/main/java/org/spongepowered/common/network/PacketUtil.java
@@ -41,7 +41,8 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.common.event.InternalNamedCauses;
 import org.spongepowered.common.event.tracking.CauseTracker;
 import org.spongepowered.common.event.tracking.PhaseContext;
-import org.spongepowered.common.event.tracking.phase.PacketPhase;
+import org.spongepowered.common.event.tracking.phase.packet.GeneralPacketState;
+import org.spongepowered.common.event.tracking.phase.packet.IPacketState;
 import org.spongepowered.common.event.tracking.phase.TrackingPhases;
 import org.spongepowered.common.interfaces.world.IMixinWorldServer;
 import org.spongepowered.common.item.inventory.util.ItemStackUtil;
@@ -97,7 +98,7 @@ public class PacketUtil {
                 final ItemStackSnapshot cursor = ItemStackUtil.snapshotOf(packetPlayer.inventory.getItemStack());
                 final IMixinWorldServer world = (IMixinWorldServer) packetPlayer.worldObj;
                 final CauseTracker causeTracker = world.getCauseTracker();
-                final PacketPhase.IPacketState packetState = TrackingPhases.PACKET.getStateForPacket(packetIn);
+                final IPacketState packetState = TrackingPhases.PACKET.getStateForPacket(packetIn);
                 if (packetState == null) {
                     throw new IllegalArgumentException("Found a null packet phase for packet: " + packetIn.getClass());
                 }
@@ -116,7 +117,7 @@ public class PacketUtil {
                     context.complete();
                     causeTracker.switchToPhase(packetState, context);
                 } else {
-                    causeTracker.switchToPhase(PacketPhase.General.INVALID, EMPTY_INVALID);
+                    causeTracker.switchToPhase(GeneralPacketState.INVALID, EMPTY_INVALID);
                 }
                 packetIn.processPacket(netHandler);
                 causeTracker.completePhase();
